CNN/requirements..txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.114.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.2.0
h11==0.14.0
idna==3.8
importlib-metadata==8.4.0
importlib-resources==6.4.5
jax==0.4.13
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mediapipe==0.10.11
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.0
numpy==1.24.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
protobuf==3.20.3
pycparser==2.22
pydantic==2.9.1
pydantic-core==2.23.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
starlette==0.38.5
sympy==1.12
threadpoolctl==3.5.0
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tqdm==4.66.5
typing-extensions==4.12.2
uvicorn==0.30.6
zipp==3.20.1
